requirements.txt
apache-airflow[google]==2.10.4
pandas==2.2.3
google-cloud-bigquery==3.27.0